[Editorial] Remove quotes from blockquote
This blockquote used non-standard quoting: it only had one opening in the first paragraph and one closing in the last paragraph, with none at all in the intervening paragraphs. Since it's not being read out loud, and since some change was going to have to be made, removing the two quotes seemed less intrusive and more proper than adding several additional opening ones on the other paragrpahs.